+++ actions/category/delete-category.ts
@@ -3,6 +3,10 @@
 import { supabaseClient } from '@/lib/supabase';
 
 export async function deleteCategory(id: number) {
+  if (process.env.NODE_ENV !== 'development') {
+    return { error: 'دسترسی محدود شده است!' };
+  }
+
   const { error } = await supabaseClient
     .from('categories')
     .delete()
